memory: split page handling into PageManager and trait PageAllocator
Things we getting a bit messy in our PageAllocator (and it didn't just
allocate pages).
The idea is, our page manager just keeps track of physical pages on our
systems (and it will link back to VMAs later on).
And it just has the functionality of being a page allocator (shown by
the trait).