refactor(mm): put the map/reload_pt methods into their own trait
Why was this done:
Before the split, if you wanted to pass the memory manager to a
function, you had the specify MemoryManger's type parameter ie:
  do_something(mm: MemoryManager<arch::MemoryImpl>)
That required exposing the underlying MemoryImpl to the user (the
function) but the whole ideo of the MemoryManager was abstracting all
that.

Now since the functionaly is exposed via a trait, a follow only needs
the follwing:
  do_something(mm: &mut dyn MemoryManager)
It doesn't know a single thing about the underlying implementation.