refactor(mm): split page allocator from memory management
The page allocator was contained inside the memory management structure.
MemoryManagement is supposed to abstract low level memory (MMU) stuff.
There wasn't a good reason for the page allocator to be with it, so I
put it appart.

There was a second reason for this, in the future we will probably
want a memory manager for each process we create, but we wouldn't have
wanted a page allocatorfor each process.