pmm: make discovery of memory regions *much* faster
Before, we had a lot of closures we would repeatadly chain to figure out which memory is what.
Now we only take into account "available" memory region, region that are not reserved,
don't have the device tree, don't have the kernel there etc...

Thanks to this the code now only iterates over memory which we is allocatable
(reading/writing to it must not break the kernel).

Co-authored-by: Esteban Skallwar Blanc <estblcsk@gmail.com>
Co-authored-by: Arthur Cohen <cohenarthur.dev@gmail.com>