put drivers/utils back into kernel
Drivers were initially put outside the kernel hoping it would with
modularity and make it easier to launch them as userland processes
later.
The issue is, our only drivers so far (UARTs and IRQ chips) are very
needed by the kernel. We criticaly *NEED* to be able to output and to handle
interrupts right in the kernel from EL1, therefore I think they should be in
the kernel (so far having theoutside the kernel is only hampering development).

But no worries, I plan on putting the other drivers in userland :D