misc: Swap staticvec with arrayvec
Staticvec is not compiling with latest nightly. As we do not really
need "static" vectors but just some kind of vectors, let's just use
arrayvec which initialize static size vectors on the stack. A bonus side
effect is that this vectors will reside on the stack instead of some
segment and thus be released after the address space is mapped.

Signed-off-by: Esteban Blanc <estblcsk@gmail.com>