add feature to compile for launching tests
Previously we were always launching the in-kernel tests.
Add a feature gates so that you don't launch tests by default.

For example, run the tests on aarch64 with the folliwng:
```
cd aarch64_qemuvirt; cargo run -F launch_tests
```

Additionaly, with this we can compile the tests to run on embedded
platforms easily.