refactor: PageAlloc trait instead of PageAllocFn
PageAllocFn was not very wieldy, the alloc_pages_* (alloc_pages_for_hal
etc...) functions were a mess and it was hard to know what to use when
etc...

Use a trait instead, that way we can still pass an allocator that does
not allocate in cases when we don't want to allocate, and pass the whole
allocator when allocations actually are needed.
That way it is easier to add functionality for when the HALs need more
info from the allocator. Also it is much much easier to manage
ownership&lifetimes this way.